Naive MidiScheduler tested

+++ jacobs_ladder/cpp/src/QpcUtils.cpp
@@ -38,7 +38,7 @@ QpcUtils::~QpcUtils() {
     SetThreadAffinityMask(GetCurrentThread(), mPreviousMask);
 }
 
-long long QpcUtils::qpcGetTicks() const {
+uint64_t QpcUtils::qpcGetTicks() const {
     LARGE_INTEGER ticks;
     if (QueryPerformanceCounter(&ticks)) {
         return ticks.QuadPart;
@@ -62,14 +62,14 @@ void QpcUtils::qpcSetFrequency(bool printMsgs) {
     }
 }
 
-void QpcUtils::qpcCoarseSleep(long long ms) {
+void QpcUtils::qpcCoarseSleep(uint64_t ms) {
     // Sleep for the specified number of milliseconds using coarse sleep (1 ms resolution since we used winmm function timeBeginPeriod(1))
     std::this_thread::sleep_for(std::chrono::milliseconds(ms));
 }
 
-void QpcUtils::qpcSleep(int option, long long dimmensionlessTime) {
-    long long dimensionlessTimeInNs;
-    long long conversionFactor;
+void QpcUtils::qpcSleep(int option, uint64_t dimmensionlessTime) {
+    uint64_t dimensionlessTimeInNs;
+    uint64_t conversionFactor;
     // If option is 0, convert the time to nanoseconds
     if (option == 0) {
         dimensionlessTimeInNs = dimmensionlessTime;
@@ -96,13 +96,13 @@ void QpcUtils::qpcSleep(int option, long long dimmensionlessTime) {
 
     if (QueryPerformanceCounter(&start)) {
         // Calculate target QPC value (absolute time)
-        long long targetTicks = start.QuadPart + (dimmensionlessTime * mFrequencyHz) / conversionFactor;
+        uint64_t targetTicks = start.QuadPart + (dimmensionlessTime * mFrequencyHz) / conversionFactor;
         
         // If the requested sleep time is greater than 2 times mPrecisionNs, use coarse sleep
-        long long threshold = 5 * mPrecisionNs;
+        uint64_t threshold = 5 * mPrecisionNs;
         if (dimensionlessTimeInNs > threshold) {
             // Coarse sleep for the requested time minus 2 times the course time resolution 
-            long long coarseSleepTime = static_cast<long long>(dimensionlessTimeInNs - threshold);
+            uint64_t coarseSleepTime = static_cast<uint64_t>(dimensionlessTimeInNs - threshold);
             std::this_thread::sleep_for(std::chrono::nanoseconds(coarseSleepTime));
         }
 
@@ -115,25 +115,25 @@ void QpcUtils::qpcSleep(int option, long long dimmensionlessTime) {
     }
 }
 
-void QpcUtils::qpcSleepNs(long long ns) {
+void QpcUtils::qpcSleepNs(uint64_t ns) {
     qpcSleep(0, ns);
 }
 
-void QpcUtils::qpcSleepUs(long long us) {
+void QpcUtils::qpcSleepUs(uint64_t us) {
     qpcSleep(1, us);
 }
 
-void QpcUtils::qpcSleepMs(long long ms) {
+void QpcUtils::qpcSleepMs(uint64_t ms) {
     qpcSleep(2, ms);
 }
 
-void QpcUtils::qpcPrintTimeDiff(int option, long long start, long long end) const {
+void QpcUtils::qpcPrintTimeDiff(int option, uint64_t start, uint64_t end) const {
     if (start < 0 || end < 0 || option < 0 || option > 2) {
         std::cerr << "Error: Invalid start or end time.\n";
         return;
     }
 
-    long long conversionFactor;
+    uint64_t conversionFactor;
     std::string units;
     // If option is 0, convert the time to nanoseconds
     if (option == 0) {
@@ -152,44 +152,44 @@ void QpcUtils::qpcPrintTimeDiff(int option, long long start, long long end) cons
     }
 
     // Calculate elapsed time in nanoseconds correctly
-    long long elapsedNs = (end - start) * conversionFactor / mFrequencyHz;
+    uint64_t elapsedNs = (end - start) * conversionFactor / mFrequencyHz;
 
     // Force decimal notation with no scientific output
     std::cout << std::fixed << std::setprecision(0)
               << "Elapsed time: " << elapsedNs << " " << units << "\n";
 }
 
-void QpcUtils::qpcPrintTimeDiffNs(long long start, long long end) const {
+void QpcUtils::qpcPrintTimeDiffNs(uint64_t start, uint64_t end) const {
     qpcPrintTimeDiff(0, start, end);
 }
 
-void QpcUtils::qpcPrintTimeDiffUs(long long start, long long end) const {
+void QpcUtils::qpcPrintTimeDiffUs(uint64_t start, uint64_t end) const {
     qpcPrintTimeDiff(1, start, end);
 }
 
-void QpcUtils::qpcPrintTimeDiffMs(long long start, long long end) const {
+void QpcUtils::qpcPrintTimeDiffMs(uint64_t start, uint64_t end) const {
     qpcPrintTimeDiff(2, start, end);
 }
 
-std::tuple<double, double, long long, double> QpcUtils::qpcDisplayStatistics(const std::vector<long long>& durations) const {
+std::tuple<double, double, uint64_t, double> QpcUtils::qpcDisplayStatistics(const std::vector<uint64_t>& durations) const {
     if (durations.empty()) {
         std::cerr << "Error: Empty duration vector passed to displayStatistics.\n";
         return std::make_tuple(0.0, 0.0, 0LL, 0.0);
     }
 
     // Convert the raw ticks to nanoseconds
-    std::vector<long long> durationsInNs;
-    for (long long duration : durations) {
-        long long durationInNs = static_cast<long long>((duration * 1.0 / mFrequencyHz) * 1e9);
+    std::vector<uint64_t> durationsInNs;
+    for (uint64_t duration : durations) {
+        uint64_t durationInNs = static_cast<uint64_t>((duration * 1.0 / mFrequencyHz) * 1e9);
         durationsInNs.push_back(durationInNs);
     }
 
     // Mean Calculation
-    long long sum = std::accumulate(durationsInNs.begin(), durationsInNs.end(), 0LL);
+    uint64_t sum = std::accumulate(durationsInNs.begin(), durationsInNs.end(), 0LL);
     double mean = static_cast<double>(sum) / durationsInNs.size();
 
     // Median Calculation
-    std::vector<long long> sortedDurations = durationsInNs;
+    std::vector<uint64_t> sortedDurations = durationsInNs;
     std::sort(sortedDurations.begin(), sortedDurations.end());
     double median;
     size_t n = sortedDurations.size();
@@ -200,12 +200,12 @@ std::tuple<double, double, long long, double> QpcUtils::qpcDisplayStatistics(con
     }
 
     // Mode Calculation
-    std::map<long long, int> frequencyMap;
-    for (long long duration : durationsInNs) {
+    std::map<uint64_t, int> frequencyMap;
+    for (uint64_t duration : durationsInNs) {
         frequencyMap[duration]++;
     }
 
-    long long mode = durationsInNs[0];
+    uint64_t mode = durationsInNs[0];
     int maxCount = 0;
     for (const auto& entry : frequencyMap) {
         if (entry.second > maxCount) {
@@ -216,7 +216,7 @@ std::tuple<double, double, long long, double> QpcUtils::qpcDisplayStatistics(con
 
     // Standard Deviation Calculation
     double variance = 0.0;
-    for (long long duration : durationsInNs) {
+    for (uint64_t duration : durationsInNs) {
         variance += std::pow(duration - mean, 2);
     }
     variance /= durationsInNs.size();
