+++ jacobs_ladder/cpp/test/QpcUtilsTest.cpp
@@ -7,15 +7,15 @@
 
 int main() {
     QpcUtils timer;
-    std::vector<long long> durations;
-    std::tuple<double, double, long long, double> stats;
+    std::vector<uint64_t> durations;
+    std::tuple<double, double, uint64_t, double> stats;
     double percentError;
 
     // Test the accuracy of sleeps when using the default sleep function (10 ms)
     for (unsigned int i = 0; i < 100; i++) {
-        long long start = timer.qpcGetTicks();
+        uint64_t start = timer.qpcGetTicks();
         std::this_thread::sleep_for(std::chrono::milliseconds(10));
-        long long end = timer.qpcGetTicks();
+        uint64_t end = timer.qpcGetTicks();
         durations.push_back(end - start);
     }
     std::cout << "Using std::this_thread_sleep_for(std::chrono::milliseconds(10))\n";
@@ -26,9 +26,9 @@ int main() {
 
     // Test the accuracy of sleeps when using the course sleep function (1 ms, this is a wrapper around std::this_sleep_for() but uses winmm function timeBeginPeriod(1) for higher accuracy, 1 ms precision)
     for (unsigned int i = 0; i < 100; i++) {
-        long long start = timer.qpcGetTicks();
+        uint64_t start = timer.qpcGetTicks();
         timer.qpcCoarseSleep(10);
-        long long end = timer.qpcGetTicks();
+        uint64_t end = timer.qpcGetTicks();
         durations.push_back(end - start);
     }
     std::cout << "Using QpcUtils::qpcCoarseSleep(10)\n";
@@ -39,9 +39,9 @@ int main() {
 
     // Test the accuracy of sleeps when sleeping for 1 ms using QpcUtils::qpcSleepMs
     for (unsigned int i = 0; i < 10000; i++) {
-        long long start = timer.qpcGetTicks();
+        uint64_t start = timer.qpcGetTicks();
         timer.qpcSleepMs(1);
-        long long end = timer.qpcGetTicks();
+        uint64_t end = timer.qpcGetTicks();
         durations.push_back(end - start);
     }
     std::cout << "Using QpcUtils::qpcSleepMs(1) i.e. 1 ms\n";
@@ -52,9 +52,9 @@ int main() {
     
     // Test accuracy of sleeps when sleeping for 33 ms using QpcUtils::qpcSleepNs (1 ms above the threshold for using course sleep)
     for (unsigned int i = 0; i < 1000; i++) {
-        long long start = timer.qpcGetTicks();
+        uint64_t start = timer.qpcGetTicks();
         timer.qpcSleepNs(33000000);
-        long long end = timer.qpcGetTicks();
+        uint64_t end = timer.qpcGetTicks();
         durations.push_back(end - start);
     }
     std::cout << "Using QpcUtils::qpcSleepNs(33000000) i.e. 33 ms\n";
@@ -65,9 +65,9 @@ int main() {
 
     // Test the accuracy of sleeps when sleeping for 1000 ms (1 second checking for good coarse time accuracy)
     for (unsigned int i = 0; i < 30; i++) {
-        long long start = timer.qpcGetTicks();
+        uint64_t start = timer.qpcGetTicks();
         timer.qpcSleepUs(1000000);
-        long long end = timer.qpcGetTicks();
+        uint64_t end = timer.qpcGetTicks();
         durations.push_back(end - start);
     }
     std::cout << "Using QpcUtils::qpcSleepUs(1000000) i.e. 1000 ms\n";
